add list of federated communities
